+++ src/Packets/EntityRelativeMovePacket.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 public struct EntityRelativeMovePacket : IPacket
 {
 	public byte ID { get { return 0x1F; } }
@@ -29,7 +31,7 @@ public struct EntityRelativeMovePacket : IPacket
 		Entity entity = BetaClient.Game.World.GetEntity(EntityID);
 		if (entity != null)
 		{
-			entity.Position += new Vector3(DeltaX / 32.0, DeltaY / 32.0, DeltaZ / 32.0);
+			entity.Position += new Vector3((float)DeltaX / 32.0f, (float)DeltaY / 32.0f, (float)DeltaZ / 32.0f);
 		}
 		else
 		{
